Buttons Have been added and a Upgrade Page
Added buttons to main page, added a upgrade page aswell, and buttons along with it.
