+++ DogeClicker/app/src/main/java/com/example/dogeclicker/models/Event.java
@@ -1,4 +1,4 @@
-package models;
+package com.example.dogeclicker.models;
 
 public class Event {
     private String name;
