+++ DogeClicker/app/src/main/java/com/example/dogeclicker/models/PC.java
@@ -1,0 +1,4 @@
+package com.example.dogeclicker.models;
+
+public class PC {
+}
