+++ DogeClicker/app/src/main/java/com/example/dogeclicker/models/UpgradeType.java
@@ -1,4 +1,4 @@
-package models;
+package com.example.dogeclicker.models;
 
 public enum UpgradeType {
     BASIC,
