+++ DogeClicker/app/src/main/java/com/example/dogeclicker/view/MainActivity.java
@@ -2,10 +2,10 @@ package com.example.dogeclicker.view;
 
 import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
+import android.view.View;
 
 import com.example.dogeclicker.R;
 
-import static java.lang.System.out;
 
 public class MainActivity extends AppCompatActivity {
 
@@ -13,5 +13,29 @@ public class MainActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
+
+    }
+
+
+
+    public void onUpgradeCilck(View v){
+        setContentView(R.layout.perm_upgrade_page);
+
+    }
+
+    public void onDogeCoinClick(View v){
+
+    }
+
+    public void onCursorClick(View v){
+
+
+    }
+
+    public void onCPUClick(View v){
+
+    }
+    public void onRAMClick(View v){
+
     }
 }
