+++ DogeClicker/app/src/main/java/com/example/dogeclicker/view/PermUpgradePage.java
@@ -1,12 +1,20 @@
 package com.example.dogeclicker.view;
 
+import android.os.Bundle;
 import android.view.View;
+import androidx.appcompat.app.AppCompatActivity;
 
-public class PermUpgradePage {
+import com.example.dogeclicker.R;
 
-    public void onCreate(){
+public class PermUpgradePage extends AppCompatActivity {
+
+
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.perm_upgrade_page);
 
     }
+
 
     public void launchActivity(){
 
@@ -24,6 +32,6 @@ public class PermUpgradePage {
 
     }
     public void onNextPcClick(View v){
-        
+
     }
 }
