--- DogeClicker/app/src/main/java/models/PC.java
@@ -1,4 +1,0 @@
-package models;
-
-public class PC {
-}
