--- DogeClicker/app/src/main/java/models/Upgrade.java
@@ -1,4 +1,0 @@
-package models;
-
-public class Upgrade {
-}
